Rename make_all_dists to main_branch_bot
oca-main-branch-bot now does all post-processing needed
on addons repo main branhces. Last missing piece in this script is
the addons table generator that we will soon replace
by a repo-level README.rst generator (from a
repo-level DESCRIPTION.rst).
