+++ setup.cfg
@@ -31,7 +31,7 @@ console_scripts =
     oca-gen-all-addons-tables = tools.gen_all_addons_tables:main
     oca-migrate-branch = tools.migrate_branch:main
     oca-migrate-branch-empty = tools.migrate_branch_empty:main
-    oca-make-all-dists = tools.make_all_dists:main
+    oca-main-branch-bot = tools.main_branch_bot:main
     oca-pypi-upload = tools.pypi_upload:cli
     oca-dist-to-simple-index = tools.dist_to_simple_index:main
     oca-gen-addon-readme = tools.gen_addon_readme:gen_addon_readme
